NMS-12731 Optimize Performance of Timeseries Integration Layer: move LateAggregationParams to its own class, move all Newts classes to the aggregation package (which is the only onw that should depend on Newts)
