+++ features/timeseries/src/main/java/org/opennms/netmgt/timeseries/integration/aggregation/NewtsLikeSampleAggregator.java
@@ -36,8 +36,11 @@ import java.time.Instant;
 import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
+import java.util.StringJoiner;
 
 import org.opennms.integration.api.v1.timeseries.Metric;
+import org.opennms.netmgt.measurements.model.Source;
+import org.opennms.netmgt.timeseries.integration.LateAggregationParams;
 import org.opennms.newts.aggregate.ResultProcessor;
 import org.opennms.newts.api.Duration;
 import org.opennms.newts.api.Measurement;
@@ -45,7 +48,9 @@ import org.opennms.newts.api.Resource;
 import org.opennms.newts.api.Results;
 import org.opennms.newts.api.Sample;
 import org.opennms.newts.api.Timestamp;
+import org.opennms.newts.api.query.AggregationFunction;
 import org.opennms.newts.api.query.ResultDescriptor;
+import org.opennms.newts.api.query.StandardAggregationFunctions;
 
 /** Aggregates the 'Newts' way. */
 public class NewtsLikeSampleAggregator {
@@ -54,16 +59,47 @@ public class NewtsLikeSampleAggregator {
     private final Timestamp start;
     private final Timestamp end;
     private final ResultDescriptor resultDescriptor;
+    private final List<Source> currentSources;
     private final Duration resolution;
     private final Metric metric;
 
-    private NewtsLikeSampleAggregator(Resource resource, Timestamp start, Timestamp end, ResultDescriptor descriptor, Duration resolution, Metric metric) {
+    private NewtsLikeSampleAggregator(Resource resource, Timestamp start, Timestamp end, List<Source> currentSources,
+                                      final LateAggregationParams lag, Metric metric) {
         this.resource = checkNotNull(resource, "resource argument");
         this.start = checkNotNull(start, "start argument");
         this.end = checkNotNull(end, "end argument");
-        resultDescriptor = checkNotNull(descriptor, "result descriptor argument");
-        this.resolution = checkNotNull(resolution, "resolution argument");
+        this.currentSources = checkNotNull(currentSources, "currentSources argument");
+        checkNotNull(lag, "lag argument");
+        this.resultDescriptor = createResultDescriptor(this.currentSources, lag);
+        this.resolution = Duration.millis(lag.getStep());
         this.metric = checkNotNull(metric, "metric argument");
+    }
+
+    private ResultDescriptor createResultDescriptor(final List<Source> listOfSources, final LateAggregationParams lag) {
+        ResultDescriptor resultDescriptor = new ResultDescriptor(lag.getInterval());
+        for (Source source : listOfSources) {
+            // Use the datasource as the metric name if set, otherwise use the name of the attribute
+            final String metricName = source.getDataSource() != null ? source.getDataSource() : source.getAttribute();
+            final String name = source.getLabel();
+            final AggregationFunction fn = toAggregationFunction(source.getAggregation());
+
+            resultDescriptor.datasource(name, metricName, lag.getHeartbeat(), fn);
+            resultDescriptor.export(name);
+        }
+        return resultDescriptor;
+    }
+
+    // Newts
+    private static AggregationFunction toAggregationFunction(String fn) {
+        if ("average".equalsIgnoreCase(fn) || "avg".equalsIgnoreCase(fn)) {
+            return org.opennms.newts.api.query.StandardAggregationFunctions.AVERAGE;
+        } else if ("max".equalsIgnoreCase(fn)) {
+            return org.opennms.newts.api.query.StandardAggregationFunctions.MAX;
+        } else if ("min".equalsIgnoreCase(fn)) {
+            return StandardAggregationFunctions.MIN;
+        } else {
+            throw new IllegalArgumentException("Unsupported aggregation function: " + fn);
+        }
     }
 
     public static NewtsLikeSampleAggregatorBuilder builder() {
@@ -87,15 +123,15 @@ public class NewtsLikeSampleAggregator {
         private Resource resource;
         private Timestamp start;
         private Timestamp end;
-        private ResultDescriptor resultDescriptor;
-        private Duration resolution;
+        private List<Source> currentSources;
+        private LateAggregationParams lateAggregationParams;
         private Metric metric;
 
         NewtsLikeSampleAggregatorBuilder() {
         }
 
-        public NewtsLikeSampleAggregatorBuilder resource(Resource resource) {
-            this.resource = resource;
+        public NewtsLikeSampleAggregatorBuilder resource(String resourceId) {
+            this.resource = new Resource(resourceId);
             return this;
         }
 
@@ -109,13 +145,13 @@ public class NewtsLikeSampleAggregator {
             return this;
         }
 
-        public NewtsLikeSampleAggregatorBuilder resultDescriptor(ResultDescriptor resultDescriptor) {
-            this.resultDescriptor = resultDescriptor;
+        public NewtsLikeSampleAggregatorBuilder currentSources(List<Source> currentSources) {
+            this.currentSources = currentSources;
             return this;
         }
 
-        public NewtsLikeSampleAggregatorBuilder resolution(Duration resolution) {
-            this.resolution = resolution;
+        public NewtsLikeSampleAggregatorBuilder lag(final LateAggregationParams lateAggregationParams) {
+            this.lateAggregationParams = lateAggregationParams;
             return this;
         }
 
@@ -125,11 +161,19 @@ public class NewtsLikeSampleAggregator {
         }
 
         public NewtsLikeSampleAggregator build() {
-            return new NewtsLikeSampleAggregator(resource, start, end, resultDescriptor, resolution, metric);
+            return new NewtsLikeSampleAggregator(resource, start, end, currentSources, lateAggregationParams, metric);
         }
 
+        @Override
         public String toString() {
-            return "NewtsLikeSampleAggregator.NewtsLikeSampleAggregatorBuilder(resource=" + this.resource + ", start=" + this.start + ", end=" + this.end + ", resultDescriptor=" + this.resultDescriptor + ", resolution=" + this.resolution + ", metric=" + this.metric + ")";
+            return new StringJoiner(", ", NewtsLikeSampleAggregatorBuilder.class.getSimpleName() + "[", "]")
+                    .add("resource=" + resource)
+                    .add("start=" + start)
+                    .add("end=" + end)
+                    .add("currentSources=" + currentSources)
+                    .add("lateAggregationParams=" + lateAggregationParams)
+                    .add("metric=" + metric)
+                    .toString();
         }
     }
 }
